Undo coffeescript-related changes made on release-1.5.2.
This will make it easier to merge devel into release-1.5.2, since devel
now contains the final verison of these changes, as implemented by
@GeoffreyBooth in #9018.

Revert "Bump coffeescript package version to 1.13.0."
This reverts commit d727ad02a71678eaee668aa37efe29ac87d05c84.

Revert "Move coffeescript and coffeescript-test-helper packages back into core. (#8960)"
This reverts commit eb3c7dd3443497ee33c04784cf5536c81867ff82.

Revert "Split coffeescript package into coffeescript / coffeescript-compiler."
This reverts commit 8344cbf59d9fa28bc16bffe6aaf8ced1256b81be.

Revert "Instructions for how to test the coffeescript package"
This reverts commit 491cbc3bfe98699be1a5ffa12500241cda989af9.
