--- packages/non-core/coffeescript-compiler/.gitignore
@@ -1,1 +1,0 @@
-.npm
