+++ packages/non-core/coffeescript/.gitignore
@@ -1,2 +1,1 @@
 .build*
-.npm
