+++ packages/non-core/coffeescript/.npm/plugin/compileCoffeescript/README
