+++ packages/non-core/coffeescript/tests/bare_test_setup.coffee
