+++ packages/non-core/coffeescript/tests/bare_tests.js
