+++ packages/non-core/coffeescript/tests/coffeescript_module.coffee
