+++ packages/non-core/coffeescript/tests/coffeescript_strict_tests.coffee
