+++ packages/non-core/coffeescript/tests/coffeescript_test_setup.js
