+++ packages/non-core/coffeescript/tests/coffeescript_tests.coffee
