+++ packages/non-core/coffeescript/tests/es2015_module.js
