+++ packages/non-core/coffeescript/tests/litcoffeescript_tests.litcoffee
